Decomposed todo filter logic to useFilter hook.
