+++ src/components/Toolbar/FilterTodo/FilterTodo.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 
-import { useActions, useAppSelector } from '@redux';
 import { FilterOptions } from '@types';
+import { useFilter } from '@hooks';
 
 import { CustomButton } from '@components';
 
@@ -10,22 +10,13 @@ import style from './FilterTodo.module.scss';
 const FILTER_OPTIONS = ['All', 'Active', 'Completed'];
 
 export const FilterTodo: FC = () => {
-  const { filterValue, todoList } = useAppSelector((state) => state.todo);
+  const {
+    filterValue,
+    isAnyTodoDone,
 
-  const { clearDoneTodo, setFilterTodo } = useActions();
-
-  const handleClearDoneTodo = () => {
-    const isConfirmed = window.confirm(
-      'Would you like to clear completed tasks?',
-    );
-
-    if (!isConfirmed) return;
-
-    clearDoneTodo();
-  };
-  const handleSetFilter = (filter: FilterOptions) => setFilterTodo(filter);
-
-  const isAnyTodoDone = !!todoList.find((todo) => todo.isDone);
+    onSetFilter,
+    onClearDoneTodos,
+  } = useFilter();
 
   return (
     <ul className={style.container}>
@@ -36,7 +27,7 @@ export const FilterTodo: FC = () => {
         >
           <CustomButton
             variant={filterValue === option ? 'default' : 'primary'}
-            onClick={() => handleSetFilter(option as FilterOptions)}
+            onClick={() => onSetFilter(option as FilterOptions)}
             isDisabled={filterValue === option}
           >
             {option}
@@ -47,7 +38,7 @@ export const FilterTodo: FC = () => {
       <li>
         <CustomButton
           variant="secondary"
-          onClick={handleClearDoneTodo}
+          onClick={onClearDoneTodos}
           isDisabled={!isAnyTodoDone}
         >
           Clear completed
