+++ src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useTodoTask';
 export * from './useModal';
+export * from './useFilter';
